- Fixed trade_traderequest which wasn't updated correctly after the @noask update. - Fixed a few compilation warnings.
git-svn-id: http://eathena.googlecode.com/svn/branches/stable@7185 9c4de8d1-4af9-ca9b-867f-b47ac08f6c80

+++ src/map/guild.c
@@ -367,7 +367,7 @@ int guild_payexp_timer_sub(DBKey dataid, void *data, va_list ap)
 		return 0;
 	}
 
-	if (g->member[i].exp > UINT_MAX - c->exp)
+	if (g->member[i].exp > UINT_MAX - (unsigned int)c->exp)
 		g->member[i].exp = UINT_MAX;
 	else
 		g->member[i].exp+= c->exp;
